Rename index.tsx -> Controller.tsx
