+++ src/ui/Player/Controller/index.tsx
@@ -1,31 +1,2 @@
-import React from "react";
-import { Audioset } from "../../../audioset";
-import { ControlState, PlayerControl } from "../../../player/Control";
 import "./Controller.css";
-import { Track } from "./Track";
-
-interface ControllerProps {
-  audioset: Audioset;
-  state: ControlState;
-  control: PlayerControl;
-}
-
-export const Controller = ({ audioset, state, control }: ControllerProps) => {
-  if (!audioset || !audioset.tracks) {
-    return <div>Audioset not loaded</div>;
-  }
-
-  return (
-    <div className="Controller">
-      {audioset.tracks.map(track => (
-        <Track
-          key={track.id}
-          track={track}
-          audioset={audioset}
-          control={control}
-          state={state}
-        />
-      ))}
-    </div>
-  );
-};
+export * from "./Controller";
